Initial Commits To Handle Some Issues
This commit should address issues #43, #42, #38, #37, #29.

+++ src/library_galilMotionController/programs/program_variable_list.h
@@ -17,7 +17,7 @@ public:
 
     bool doesVariableExist(const std::string &name) const;
 
-    void addVariable(const std::string &name, const int &lineNumber);
+    void addVariable(const std::string &name, const double &lineNumber);
 
     void removeVariable(const std::string &name);
